Refactor the ping command to use threads
Oh the irony of writing a plugin to monitor server lag, then add a
command that blocks the game thread for up to 10 seconds, while waiting
on a "while" loop for an external process to exit. And... this causes lag!

Bonus points for using the word "exponentially" to describe the time
complexity of an operation that involves sending a constant number of
packets and getting the same number of replies.

I removed the 10 packet limitation and changed it with a warning saying
that it might take a long time. Feel free to change this part to
something else - since this command uses nearly 0 cpu time, I felt it
made no sense to block this feature.

(Heh, "block")